added our populate teams module to our build script previously our teams kept disappearing when we forgot to run the populate script after the build. Makes sense to combine them... Also got rid of annoying console.logs in a number of places including the team section relates to #69
